requirements.txt
streamlit
pandas
yfinance
plotly
requests
psycopg2-binary